requirements.txt
fastapi
uvicorn
pillow
insightface
onnxruntime
opencv-python
python-dotenv